Use static routing for /notebook
