+++ next.config.ts
@@ -1,7 +1,10 @@
 import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {
-  /* config options here */
+  output: 'export',
+  images: { unoptimized: true },
+  trailingSlash: false,
+  generateBuildId: async () => 'static-build'
 };
 
 export default nextConfig;
